Munge the build version to make the App Store happy.
Apple rejected 3.1.1 on the grounds that the bundle version of 2025.06.19 was lower
than the old one, of 2026. At first I thought I must have mistyped the year, but it
looks like it just doesn't like dotted numbers in that field and could not compare
the two. Workaround - remove the dots! The App Store accepted 20250619 as higher.